Improved update.sh behavior for redcap_cypress

+++ update.sh
@@ -3,8 +3,6 @@
 # The following line is important to ensure this script stops if any repos have outstanding changes that would prevent a git pull.
 set -e
 
-set -x
-
 getUpdateScriptInfo () {
   ls -l update.sh
 }
@@ -31,6 +29,7 @@ if [ "$rsvcBranchName" = "staging" ] || [ "$rsvcBranchName" = "main" ]; then
 else
     commitsBehindStaging=`git log --oneline ..FETCH_HEAD | wc -l`
     if [ $commitsBehindStaging != 0 ]; then
+        set +x
         echo
         echo Please merge the latest from the 'staging' branch into your redcap_rsvc branch.
         echo This is not performed automatically to avoid interfering with any active development. 
@@ -40,8 +39,15 @@ fi
 cd ../..
 
 cd redcap_cypress
-git checkout master
 git pull
+git fetch https://github.com/vanderbilt-redcap/redcap_cypress master
+commitsBehindMaster=`git log --oneline ..FETCH_HEAD | wc -l`
+if [ $commitsBehindMaster != 0 ]; then
+    echo
+    echo Please either checkout the master branch for redcap_cypress, or merge it into your working branch.
+    exit
+fi
+
 # Adam Lewis had an instance where cypress started throwing confusing errors on every feature which was resolved by the following steps:
 rm node_modules -r
 npm cache clean --force
@@ -53,3 +59,5 @@ git checkout main
 git pull
 docker compose --profile external-storage --profile sftp down # This ensures a running container is restarted, which can fix various docker issues.
 docker compose up -d --build --remove-orphans # This ensures the container is rebuilt to include any Dockerfile changes, other updates, or fix various issues.
+
+echo 'Update completed successfully!'
